Create a dedicated file for the Node.js API
